+++ ARKit+CoreLocation/AppDelegate.swift
@@ -29,12 +29,8 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         UIApplication.shared.isIdleTimerDisabled = true
 
         self.window = UIWindow(frame: UIScreen.main.bounds)
-
-        self.window!.makeKeyAndVisible()
-
-        let vc = ViewController()
-
-        self.window?.rootViewController = vc
+        self.window?.makeKeyAndVisible()
+        self.window?.rootViewController = ViewController()
 
         return true
     }
