+++ ARKit+CoreLocation/Source/SceneLocationView.swift
@@ -61,7 +61,7 @@ public class SceneLocationView: ARSCNView {
 
     private var sceneLocationEstimates = [SceneLocationEstimate]()
 
-    public private(set) var sceneNode: SCNNode? {
+    public internal(set) var sceneNode: SCNNode? {
         didSet {
             guard sceneNode != nil else { return }
 
@@ -73,7 +73,7 @@ public class SceneLocationView: ARSCNView {
 
     private var updateEstimatesTimer: Timer?
 
-    private var didFetchInitialLocation = false
+    internal var didFetchInitialLocation = false
 
     ///Whether debugging feature points should be displayed.
     ///Defaults to false
@@ -90,7 +90,7 @@ public class SceneLocationView: ARSCNView {
 
     // MARK: Setup
     public convenience init() {
-        self.init(frame: CGRect.zero, options: nil)
+        self.init(frame: .zero, options: nil)
     }
 
     public override init(frame: CGRect, options: [String : Any]? = nil) {
@@ -116,12 +116,7 @@ public class SceneLocationView: ARSCNView {
         // Create a session configuration
         let configuration = ARWorldTrackingConfiguration()
         configuration.planeDetection = .horizontal
-
-        if orientToTrueNorth {
-            configuration.worldAlignment = .gravityAndHeading
-        } else {
-            configuration.worldAlignment = .gravity
-        }
+        configuration.worldAlignment = orientToTrueNorth ? .gravityAndHeading : .gravity
 
         // Run the view's session
         session.run(configuration)
@@ -185,9 +180,9 @@ public class SceneLocationView: ARSCNView {
     }
 
     ///Adds a scene location estimate based on current time, camera position and location from location manager
-    fileprivate func addSceneLocationEstimate(location: CLLocation) {
+    internal func addSceneLocationEstimate(location: CLLocation) {
         guard let position = currentScenePosition() else { return }
-        
+
         let sceneLocationEstimate = SceneLocationEstimate(location: location, position: position)
         self.sceneLocationEstimates.append(sceneLocationEstimate)
 
@@ -237,9 +232,7 @@ public class SceneLocationView: ARSCNView {
     }
 
     public func currentLocation() -> CLLocation? {
-        if locationEstimateMethod == .coreLocationDataOnly {
-            return locationManager.currentLocation
-        }
+        if locationEstimateMethod == .coreLocationDataOnly { return locationManager.currentLocation }
 
         guard let bestEstimate = self.bestLocationEstimate(),
             let position = currentScenePosition() else {
@@ -361,27 +354,18 @@ public class SceneLocationView: ARSCNView {
 
                 adjustedDistance = distance * Double(scale)
 
-                let adjustedTranslation = SCNVector3(
-                    x: Float(locationTranslation.longitudeTranslation) * scale,
-                    y: Float(locationTranslation.altitudeTranslation) * scale,
-                    z: Float(locationTranslation.latitudeTranslation) * scale)
-
-                let position = SCNVector3(
-                    x: currentPosition.x + adjustedTranslation.x,
-                    y: currentPosition.y + adjustedTranslation.y,
-                    z: currentPosition.z - adjustedTranslation.z)
-
-                locationNode.position = position
-
+                let adjustedTranslation = SCNVector3( x: Float(locationTranslation.longitudeTranslation) * scale,
+                                                      y: Float(locationTranslation.altitudeTranslation) * scale,
+                                                      z: Float(locationTranslation.latitudeTranslation) * scale)
+                locationNode.position = SCNVector3( x: currentPosition.x + adjustedTranslation.x,
+                                                    y: currentPosition.y + adjustedTranslation.y,
+                                                    z: currentPosition.z - adjustedTranslation.z)
                 locationNode.scale = SCNVector3(x: scale, y: scale, z: scale)
             } else {
                 adjustedDistance = distance
-                let position = SCNVector3(
-                    x: currentPosition.x + Float(locationTranslation.longitudeTranslation),
-                    y: currentPosition.y + Float(locationTranslation.altitudeTranslation),
-                    z: currentPosition.z - Float(locationTranslation.latitudeTranslation))
-
-                locationNode.position = position
+                locationNode.position = SCNVector3( x: currentPosition.x + Float(locationTranslation.longitudeTranslation),
+                                                    y: currentPosition.y + Float(locationTranslation.altitudeTranslation),
+                                                    z: currentPosition.z - Float(locationTranslation.latitudeTranslation))
                 locationNode.scale = SCNVector3(x: 1, y: 1, z: 1)
             }
         } else {
@@ -406,9 +390,7 @@ public class SceneLocationView: ARSCNView {
                 //Scale it to be an appropriate size so that it can be seen
                 scale = Float(adjustedDistance) * 0.181
 
-                if distance > 3_000 {
-                    scale *=  0.75
-                }
+                if distance > 3_000 { scale *=  0.75 }
 
                 annotationNode.annotationNode.scale = SCNVector3(x: scale, y: scale, z: scale)
             }
@@ -420,69 +402,5 @@ public class SceneLocationView: ARSCNView {
 
         locationDelegate?.sceneLocationViewDidUpdateLocationAndScaleOfLocationNode(sceneLocationView: self,
                                                                                    locationNode: locationNode)
-    }
-}
-
-extension SceneLocationView: ARSCNViewDelegate {
-    public func renderer(_ renderer: SCNSceneRenderer, didRenderScene scene: SCNScene, atTime time: TimeInterval) {
-        if sceneNode == nil {
-            sceneNode = SCNNode()
-            scene.rootNode.addChildNode(sceneNode!)
-
-            if showAxesNode {
-                let axesNode = SCNNode.axesNode(quiverLength: 0.1, quiverThickness: 0.5)
-                sceneNode?.addChildNode(axesNode)
-            }
-        }
-
-        if !didFetchInitialLocation {
-            //Current frame and current location are required for this to be successful
-            if session.currentFrame != nil,
-                let currentLocation = self.locationManager.currentLocation {
-                didFetchInitialLocation = true
-
-                self.addSceneLocationEstimate(location: currentLocation)
-            }
-        }
-    }
-
-    public func sessionWasInterrupted(_ session: ARSession) {
-        print("session was interrupted")
-    }
-
-    public func sessionInterruptionEnded(_ session: ARSession) {
-        print("session interruption ended")
-    }
-
-    public func session(_ session: ARSession, didFailWithError error: Error) {
-        print("session did fail with error: \(error)")
-    }
-
-    public func session(_ session: ARSession, cameraDidChangeTrackingState camera: ARCamera) {
-        switch camera.trackingState {
-        case .limited(.insufficientFeatures):
-            print("camera did change tracking state: limited, insufficient features")
-        case .limited(.excessiveMotion):
-            print("camera did change tracking state: limited, excessive motion")
-        case .limited(.initializing):
-            print("camera did change tracking state: limited, initializing")
-        case .normal:
-            print("camera did change tracking state: normal")
-        case .notAvailable:
-            print("camera did change tracking state: not available")
-        }
-    }
-}
-
-// MARK: LocationManager
-extension SceneLocationView: LocationManagerDelegate {
-    func locationManagerDidUpdateLocation(_ locationManager: LocationManager, location: CLLocation) {
-        addSceneLocationEstimate(location: location)
-    }
-
-    func locationManagerDidUpdateHeading(_ locationManager: LocationManager,
-                                         heading: CLLocationDirection,
-                                         accuracy: CLLocationAccuracy) {
-
     }
 }
